+++ module_5_wifi/module_5_wifi.ino
@@ -1,11 +1,9 @@
-#include "types.h"
 #include "TinyGPS++.h"
 #include <SoftwareSerial.h>
-#include <Wire.h>
+
 #include "EEPROM.h"
 #include <WiFi.h>
 #include <WiFiClient.h>
-#include <WiFiAP.h>
 
 // Set these to your desired credentials.
 const char *ssid = "ESP32";
@@ -30,7 +28,7 @@ int count = 0; //to take treat 1 out of Nb
 int Nb = 10; //for frequency
 bool looped = false; //to increment
 
-float memory_addr; //current memory pointer
+int memory_addr; //current memory pointer
 float longitude; //current gps longitude
 float latitude; //current gps latitude
